refactor(minigame): split livewire by game-phase seam (#244)
livewire/mod.rs was 710 lines (over the 700-line cap). Per the issue's
"seam by game phase" guidance, split the construction half of the file
out:

  livewire/mod.rs   — Wire struct, LivewireGame struct + new(),
                       MinigameInstance impl (started/message/tick/aborted/
                       needs_tick) — the runtime-loop side
  livewire/setup.rs — DifficultyParams + DIFFICULTY_LEVELS table,
                       sound/wire/electrode/location constants,
                       impl LivewireGame { init_game, update_difficulty,
                       setup_wires, build_full_game_state, find_*, random_* }
  livewire/tests.rs — unchanged

Public surface preserved: only `LivewireGame` is referenced externally
(by `games/mod.rs`); the type still lives in `mod.rs` and is `pub`.

Tests pass; clippy clean; cargo fmt applied.
